WIP:  Started extracting binding action into one super class
