+++ app/actions/service_route_binding_create.rb
@@ -1,9 +1,11 @@
 require 'services/service_brokers/service_client_provider'
+require 'actions/v3/service_binding_create'
 
 module VCAP::CloudController
   module V3
-    class ServiceRouteBindingCreate
+    class ServiceRouteBindingCreate < V3::ServiceBindingCreate
       def initialize(service_event_repository)
+        super()
         @service_event_repository = service_event_repository
       end
 
@@ -22,76 +24,9 @@ module VCAP::CloudController
         )
       end
 
-      def bind(precursor, parameters: {}, accepts_incomplete: false)
-        client = VCAP::Services::ServiceClientProvider.provide(instance: precursor.service_instance)
-        details = client.bind(precursor, arbitrary_parameters: parameters, accepts_incomplete: accepts_incomplete)
-
-        if details[:async]
-          not_retrievable! unless bindings_retrievable?(precursor)
-          save_incomplete_binding(precursor, details[:operation])
-        else
-          complete_binding_and_save(precursor, details[:binding][:route_service_url])
-        end
-      rescue => e
-        precursor.save_with_new_operation({}, {
-          type: 'create',
-          state: 'failed',
-          description: e.message,
-        })
-
-        raise e
-      end
-
-      def poll(binding)
-        client = VCAP::Services::ServiceClientProvider.provide(instance: binding.service_instance)
-
-        details = fetch_last_operation(client, binding)
-        return { finished: false } unless details
-
-        attributes = {}
-
-        complete = details[:last_operation][:state] == 'succeeded'
-        if complete
-          params = client.fetch_service_binding(binding)
-          attributes[:route_service_url] = params[:route_service_url]
-        end
-
-        binding.save_with_new_operation(
-          attributes,
-          {
-            type: 'create',
-            state: details[:last_operation][:state],
-            description: details[:last_operation][:description],
-          }
-        )
-
-        if complete
-          binding.notify_diego
-          record_audit_event(binding)
-        end
-
-        if binding.reload.terminal_state?
-          { finished: true }
-        else
-          { finished: false, retry_after: details[:retry_after] }
-        end
-      rescue => e
-        binding.save_with_new_operation({}, {
-          type: 'create',
-          state: 'failed',
-          description: e.message,
-        })
-        { finished: true }
-      end
-
-      # PollingComplete = Class.new.freeze
-      # PollingNotComplete = Struct.new(:retry_after).freeze
-
       class UnprocessableCreate < StandardError; end
 
       class RouteBindingAlreadyExists < StandardError; end
-
-      class BindingNotRetrievable < StandardError; end
 
       private
 
@@ -107,46 +42,18 @@ module VCAP::CloudController
 
       attr_reader :service_event_repository
 
-      def fetch_last_operation(client, binding)
-        client.fetch_service_binding_last_operation(binding)
-      rescue VCAP::Services::ServiceBrokers::V2::Errors::ServiceBrokerBadResponse,
-             VCAP::Services::ServiceBrokers::V2::Errors::ServiceBrokerRequestRejected,
-             HttpRequestError => e
-        binding.save_with_new_operation({}, {
-          type: 'create',
-          state: 'in progress',
-          description: e.message,
-        })
-
-        return nil
-      end
-
-      def save_incomplete_binding(precursor, operation)
-        precursor.save_with_new_operation({},
+      def complete_binding_and_save(binding, details)
+        binding.save_with_new_operation(
           {
-            type: 'create',
-            state: 'in progress',
-            broker_provided_operation: operation
-          }
-        )
-      end
-
-      def complete_binding_and_save(precursor, route_service_url)
-        save_with_route_service_url(precursor, route_service_url)
-        precursor.notify_diego
-        record_audit_event(precursor)
-      end
-
-      def save_with_route_service_url(precursor, route_service_url)
-        precursor.save_with_new_operation(
-          {
-            route_service_url: route_service_url
+            route_service_url: details[:binding][:route_service_url]
           },
           {
             type: 'create',
             state: 'succeeded',
           }
         )
+        binding.notify_diego
+        record_audit_event(binding)
       end
 
       def record_audit_event(precursor)
@@ -155,10 +62,6 @@ module VCAP::CloudController
           precursor.service_instance,
           { route_guid: precursor.route.guid },
         )
-      end
-
-      def bindings_retrievable?(binding)
-        binding.service_instance.service.bindings_retrievable
       end
 
       def operation_in_progress!
@@ -187,10 +90,6 @@ module VCAP::CloudController
 
       def already_exists!
         raise RouteBindingAlreadyExists.new('The route and service instance are already bound')
-      end
-
-      def not_retrievable!
-        raise BindingNotRetrievable.new('The broker responded asynchronously but does not support fetching binding data')
       end
     end
   end
